+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 7.0 (unreleased)
 ----------------
 
+- In the control panel you can choose to search the catalog instead of waking up every object in the path.
+
 - Removed ``ISearchReplaceable`` behavior.  This was introduced in version 6.
   Kept the interface for backwards compatibility, but it is not used anymore.
   Instead, by default all types are searched and replaced.